[#58] DisplayDayLog, RecordDetail 파일 분리
